[#5] Renamed AtackComponent files
Removed pch include, fixed other includes
