+++ Roguelike/Roguelike.vcxproj
@@ -154,7 +154,7 @@
     <ClCompile Include="ActorRegisterSystem.cpp" />
     <ClCompile Include="ArmorBarComponent.cpp" />
     <ClCompile Include="ArmorComponent.cpp" />
-    <ClCompile Include="AtackComponent.cpp" />
+    <ClCompile Include="AttackComponent.cpp" />
     <ClCompile Include="BarComponent.cpp" />
     <ClCompile Include="HealthBarComponent.cpp" />
     <ClCompile Include="HealthComponent.cpp" />
@@ -163,7 +163,7 @@
     <ClCompile Include="PerceptionComponentDebugDraw.cpp" />
     <ClCompile Include="PerceptionSystem.cpp" />
     <ClCompile Include="SpriteOpaqueComponent.cpp" />
-    <ClCompile Include="AIAtackComponent.cpp" />
+    <ClCompile Include="AIAttackComponent.cpp" />
     <ClCompile Include="AIChaseTargetComponent.cpp" />
     <ClCompile Include="AIBlackboard.cpp" />
     <ClCompile Include="AIPerceptionComponent.cpp" />
@@ -178,7 +178,7 @@
     <ClCompile Include="LabyrinthBuilder.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="Player.cpp" />
-    <ClCompile Include="PlayerAtackComponent.cpp" />
+    <ClCompile Include="PlayerAttackComponent.cpp" />
     <ClCompile Include="Settings.cpp" />
     <ClCompile Include="Wall.cpp" />
   </ItemGroup>
@@ -187,7 +187,7 @@
     <ClInclude Include="ActorRegisterSystem.h" />
     <ClInclude Include="ArmorBarComponent.h" />
     <ClInclude Include="ArmorComponent.h" />
-    <ClInclude Include="AtackComponent.h" />
+    <ClInclude Include="AttackComponent.h" />
     <ClInclude Include="BarComponent.h" />
     <ClInclude Include="HealthBarComponent.h" />
     <ClInclude Include="HealthComponent.h" />
@@ -196,7 +196,7 @@
     <ClInclude Include="PerceptionComponentDebugDraw.h" />
     <ClInclude Include="PerceptionSystem.h" />
     <ClInclude Include="SpriteOpaqueComponent.h" />
-    <ClInclude Include="AIAtackComponent.h" />
+    <ClInclude Include="AIAttackComponent.h" />
     <ClInclude Include="AIChaseTargetComponent.h" />
     <ClInclude Include="AIBlackboard.h" />
     <ClInclude Include="AIPerceptionComponent.h" />
@@ -211,7 +211,7 @@
     <ClInclude Include="Labyrinth.h" />
     <ClInclude Include="LabyrinthBuilder.h" />
     <ClInclude Include="Player.h" />
-    <ClInclude Include="PlayerAtackComponent.h" />
+    <ClInclude Include="PlayerAttackComponent.h" />
     <ClInclude Include="Randomizer.h" />
     <ClInclude Include="Settings.h" />
     <ClInclude Include="Utility.h" />
